Extract HDR burst builder into camera.request
