+++ asg_client/app/src/test/java/com/mentra/asg_client/camera/request/HdrBurstBuilderTest.java
@@ -1,4 +1,4 @@
-package com.mentra.asg_client.camera;
+package com.mentra.asg_client.camera.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -15,7 +15,7 @@ import org.robolectric.annotation.Config;
 
 /**
  * Phase 2d unit tests for {@link HdrBurstBuilder}. Verifies the bracket configuration recipe
- * matches the inline code that previously lived in {@link CameraNeo#captureHdrBurst()}.
+ * matches the inline code that previously lived in {@link CameraNeoService#captureHdrBurst()}.
  */
 @RunWith(RobolectricTestRunner.class)
 @Config(sdk = 33)
